requirements.txt
paho-mqtt
pymongo
influxdb-client